Move the items from scorecard_step_constant.js to scorecard_constant.js
