+++ app/constants/scorecard_constant.js
@@ -1,5 +1,3 @@
-import { INDICATOR_BASE_STEP, PARTICIPANT_BASE_STEP } from './scorecard_step_constant';
-
 // Download phases of scorecard contain indicator, language_indicator, caf,
 // rating_scale, program_language, lang_indicator_audio, lang_rating_scale_audio, and indicator_image
 const scorecardDownloadSteps = 6;
@@ -38,10 +36,35 @@ const scorecardTypes = [
 const INDICATOR_BASE = 'indicatorBase';
 const PARTICIPANT_BASE = 'participantBase';
 
+const INDICATOR_BASE_STEP = 11;
+const PARTICIPANT_BASE_STEP = 12;
+
 const  PROPOSED_INDICATOR_METHODS = {
   'participantBase': { id: 1, name: PARTICIPANT_BASE, step: PARTICIPANT_BASE_STEP },
   'indicatorBase': { id: 2, name: INDICATOR_BASE, step: INDICATOR_BASE_STEP },
 }
+
+const SETUP = 1;
+const PROPOSED_CRITERIA = 2;
+const INDICATOR_DEVELOPMENT = 3;
+const VOTING = 4;
+const SCORECARD_RESULT = 5;
+
+const scorecardTrackingSteps = {
+  0: 'scorecard_joined',
+  1: 'scorecard_downloaded',
+  2: 'finished_preference',
+  3: 'finished_facilitator',
+  4: 'finished_participant_list',
+  5: 'finished_proposed_criteria',
+  6: 'finished_indicator_development',
+  7: 'finished_voting',
+  8: 'finished_scorecard_result',
+  9: 'scorecard_completed',
+  10: 'scorecard_submitted',
+  11: 'propose_by_indicator_base',
+  12: 'propose_by_participant_base',
+};
 
 export {
   scorecardDownloadPhases,
@@ -59,5 +82,13 @@ export {
   scorecardTypes,
   INDICATOR_BASE,
   PARTICIPANT_BASE,
+  INDICATOR_BASE_STEP,
+  PARTICIPANT_BASE_STEP,
   PROPOSED_INDICATOR_METHODS,
+  SETUP,
+  PROPOSED_CRITERIA,
+  INDICATOR_DEVELOPMENT,
+  VOTING,
+  SCORECARD_RESULT,
+  scorecardTrackingSteps,
 };
