+++ app/services/scorecard_progress_service.js
@@ -1,5 +1,5 @@
 import AsyncStorage from '@react-native-community/async-storage';
-import { VOTING } from '../constants/scorecard_step_constant';
+import { VOTING } from '../constants/scorecard_constant';
 import VotingIndicator from '../models/VotingIndicator';
 import Scorecard from '../models/Scorecard';
